Move MC classes to their own package
