+++ stripe/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:theme="@style/StripeDefaultTheme" />
 
         <activity
-            android:name=".view.CheckoutActivity"
+            android:name=".checkout.CheckoutActivity"
             android:theme="@style/StripeCheckoutDefaultTheme" />
     </application>
