+++ stripe/res/layout/fragment_checkout_payment_methods_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.CheckoutPaymentMethodsListFragment">
+    tools:context=".checkout.CheckoutPaymentMethodsListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler"
