+++ stripe/src/main/java/com/stripe/android/checkout/CheckoutActivityStarter.kt
@@ -1,9 +1,8 @@
-package com.stripe.android
+package com.stripe.android.checkout
 
 import android.app.Activity
 import android.content.Intent
 import com.stripe.android.view.ActivityStarter
-import com.stripe.android.view.CheckoutActivity
 import kotlinx.android.parcel.Parcelize
 
 internal class CheckoutActivityStarter : ActivityStarter<CheckoutActivity, CheckoutActivityStarter.Args> {
